refactor: Use global required validator for question
The KBV CRI mostly displays dynamic questions. This change sets a
default validation message for all required fields. The only other field
in this CRI is on the abandon page, and that already has a required
message set.

This approach of overriding the default will allow for individual
questions to have their own validation message if needed.